# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CloneHallRequest:
    properties:
      name:
        example: Phong 3
        maxLength: 255
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse:
    properties:
      active:
        type: boolean
      aisle_after_cols:
        items:
          type: integer
        type: array
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      rows:
        type: integer
      screen_position:
        type: string
      seats_per_row:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallTemplateResponse:
    properties:
      name:
        type: string
      rows:
        type: integer
      seat_count:
        type: integer
      seat_count_by_type:
        additionalProperties:
          type: integer
        type: object
      seats_per_row:
        type: integer
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdateHallRequest:
    properties:
      active:
        type: boolean
      aisle_after_cols:
        items:
          type: integer
        maxItems: 49
        type: array
      name:
        maxLength: 255
        minLength: 1
        type: string
      screen_position:
        enum:
        - front
        - back
        type: string
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body:
    properties:
      code:
        example: 0
        type: integer
      data: {}
      details:
        additionalProperties:
          type: string
        type: object
      message:
        example: success
        type: string
      reason:
        description: |-
          Reason is a stable machine-readable key for an error ("discount_expired"), so the
          frontend can translate without matching on the English sentence. Absent on success.
        type: string
    type: object
...
  github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged:
    properties:
      items: {}
      meta:
        $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Meta'
    type: object
...
info:
  contact: {}
  description: API for the Cinema Project box-office system.
  title: BackEnd-CP API
  version: "1.0"
paths:
...
  /admin/hall-templates:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallTemplateResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Preview the built-in hall templates
      tags:
      - halls
  /admin/halls:
    get:
      parameters:
      - default: 1
        description: Current page
        in: query
        name: page
        type: integer
      - default: 10
        description: Records per page
        in: query
        name: page_size
        type: integer
      - description: Search by hall name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Paged'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse'
                        type: array
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: List halls (paginated)
      tags:
      - halls
    post:
      consumes:
      - application/json
      parameters:
      - description: Hall layout
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Create a hall and generate its seat grid
      tags:
      - halls
  /admin/halls/{id}:
    delete:
      description: Refused (409) while it has a showtime not yet ended.
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Delete a hall (soft delete)
      tags:
      - halls
    get:
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Get hall details
      tags:
      - halls
    put:
      consumes:
      - application/json
      description: Deactivating is refused (409) while an open showtime is still to
        come; once inactive the hall takes no new showtimes.
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.UpdateHallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Rename a hall, change its screen/aisle display, or (de)activate it
      tags:
      - halls
  /admin/halls/{id}/clone:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hall ID to clone
        in: path
        name: id
        required: true
        type: string
      - description: New hall name
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.CloneHallRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Clone a hall's current seat grid under a new name
      tags:
      - halls
  /admin/halls/{id}/layout:
    put:
      consumes:
      - application/json
      description: Only while the hall has never had a single booking (409 otherwise).
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      - description: New layout (name/prices are ignored here)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.HallResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Regenerate a hall's whole seat grid
      tags:
      - halls
  /admin/halls/{id}/seats:
    get:
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Get the seat grid of a hall
      tags:
      - halls
    patch:
      consumes:
      - application/json
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      - description: Changes, one selector each
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.BulkSeatUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Change many seats at once (labels, rows, columns or a range)
      tags:
      - halls
  /admin/halls/{id}/seats/{seatId}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      - description: Seat ID
        in: path
        name: seatId
        required: true
        type: string
      - description: Seat changes
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Change a seat's type or gap flag
      tags:
      - halls
  /admin/halls/{id}/seats/columns:
    post:
      description: Never touches an existing seat - safe while the hall has a live
        booking (409 only on an actual conflict), unlike regenerating the whole layout.
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Append one new standard seat to every existing row, widening the hall
        by one column
      tags:
      - halls
  /admin/halls/{id}/seats/merge:
    post:
      consumes:
      - application/json
      description: The right seat is deleted entirely; the left seat becomes col_span=2,
        seat_type=couple. Refused (409) if either seat has booking history.
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      - description: Left/right seat labels
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.MergeSeatsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Merge two adjacent standard seats into one couple seat
      tags:
      - halls
  /admin/halls/{id}/seats/rows:
    post:
      description: Never touches an existing seat - safe while the hall has a live
        booking (409 only on an actual conflict), unlike regenerating the whole layout.
      parameters:
      - description: Hall ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_internal_dto.SeatResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Cinema-Project-Juann_BackEnd-CP_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: Append one new row of standard seats to a hall
      tags:
      - halls
...
securityDefinitions:
  BearerAuth:
    description: 'Header format: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"